+++ src/SpeakEasy.Specifications/ParameterSpecification.cs
@@ -1,5 +1,6 @@
 using System;
 using Machine.Specifications;
+using SpeakEasy.ArrayFormatters;
 
 namespace SpeakEasy.Specifications
 {
