+++ src/SpeakEasy.Specifications/Requests/PostRequestSpecification.cs
@@ -1,5 +1,6 @@
 ﻿using Machine.Fakes;
 using Machine.Specifications;
+using SpeakEasy.ArrayFormatters;
 using SpeakEasy.Requests;
 
 namespace SpeakEasy.Specifications.Requests
