+++ src/SpeakEasy.Specifications/ResourceSpecification.cs
@@ -1,4 +1,5 @@
 ﻿using Machine.Specifications;
+using SpeakEasy.ArrayFormatters;
 
 namespace SpeakEasy.Specifications
 {
