+++ src/SpeakEasy/ArrayFormatters/MultipleValuesArrayFormatter.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Linq;
 
-namespace SpeakEasy
+namespace SpeakEasy.ArrayFormatters
 {
     public class MultipleValuesArrayFormatter : IArrayFormatter
     {
